common coadd. Initialisation functions consolidated in initialisation module
