+++ scarlet/source.py
@@ -1,345 +1,15 @@
 from functools import partial
 
+from .initialisation import *
 from .constraint import PositivityConstraint, MonotonicityConstraint, SymmetryConstraint, L0Constraint
 from .constraint import NormalizationConstraint, ConstraintChain, CenterOnConstraint
 from .parameter import Parameter, relative_step
 from .component import ComponentTree, FunctionComponent, FactorizedComponent
 from .bbox import Box
 from .wavelet import Starlet, mad_wavelet
-from .observation import Observation
-from .interpolation import interpolate_observation
-from . import operator
 
 # make sure that import * above doesn't import its own stock numpy
 import autograd.numpy as np
-
-import logging
-
-logger = logging.getLogger("scarlet.source")
-
-
-def get_pixel_sed(sky_coord, observation):
-    """Get the SED at `sky_coord` in `observation`
-
-    Parameters
-    ----------
-    sky_coord: tuple
-        Position in the observation
-    observation: `~scarlet.Observation`
-        Observation to extract SED from.
-
-    Returns
-    -------
-    SED: `~numpy.array`
-    """
-
-    pixel = observation.frame.get_pixel(sky_coord)
-    sed = observation.images[:, pixel[0], pixel[1]].copy()
-    return sed
-
-
-def get_psf_sed(sky_coord, observation, frame):
-    """Get SED for a point source at `sky_coord` in `observation`
-
-    Identical to `get_pixel_sed`, but corrects for the different
-    peak values of the observed seds to approximately correct for PSF
-    width variations between channels.
-
-    Parameters
-    ----------
-    sky_coord: tuple
-        Position in the observation
-    observation: `~scarlet.Observation`
-        Observation to extract SED from.
-    frame: `~scarlet.Frame`
-        Frame of the model
-
-    Returns
-    -------
-    SED: `~numpy.array`
-    """
-    sed = get_pixel_sed(sky_coord, observation)
-
-    # approx. correct PSF width variations from SED by normalizing heights
-    if observation._diff_kernels is not None:
-        sed /= observation._diff_kernels.image.sum(axis=(-2,-1))
-
-        return sed
-
-    if observation.frame.psf is not None:
-        # Account for the PSF in the intensity
-        sed /= observation.frame.psf.image.sum(axis=(-2, -1))
-
-    if frame.psf is not None:
-        sed = sed * frame.psf.image[0].sum()
-
-    return sed
-
-def build_sed_coadd(seds, bg_rmses, observations):
-    """Build a channel weighted coadd to use for source detection
-    Parameters
-    ----------
-    sed: array
-        SED at the center of the source.
-    bg_rms: array
-        Background RMS in each channel in observation.
-    observations: list of `~scarlet.observation.Observation`
-        Observations to use for the coadd.
-    Returns
-    -------
-    detect: array
-        2D image created by weighting all of the channels by SED
-    bg_cutoff: float
-        The minimum value in `detect` to include in detection.
-    """
-    # The observation that lives in the same plane as the frame
-    loc = np.where([type(obs) is Observation for obs in observations])
-    # If more than one element is an `Observation`, then pick the first one as a reference (arbitrary)
-    obs_ref = observations[loc[0][0]]
-    positive_img = []
-    positive_bgrms = []
-    weights = []
-    jacobian_args = []
-    for i,obs in enumerate(observations):
-        sed = seds[i]
-        C = len(seds[i])
-        bg_rms = bg_rmses[i]
-        if np.any(bg_rms <= 0):
-            raise ValueError("bg_rms must be greater than zero in all channels")
-
-        positive = [c for c in range(C) if sed[c] > 0]
-        if type(obs) is Observation:
-            positive_img += [obs.images[c] for c in positive]
-        else:
-            positive_img += [interpolate_observation(obs, obs_ref.frame)[c] for c in positive]
-        positive_bgrms += [bg_rms[c] for c in positive]
-        weights += [sed[c] / bg_rms[c] ** 2 for c in positive]
-        jacobian_args += [sed[c] ** 2 / bg_rms[c] ** 2 for c in positive]
-
-    detect = np.einsum("i,i...", np.array(weights), positive_img) / np.sum(jacobian_args)
-
-    # thresh is multiple above the rms of detect (weighted variance across channels)
-    bg_cutoff = np.sqrt((np.array(weights) ** 2 * np.array(positive_bgrms) ** 2).sum()) / np.sum(jacobian_args)
-    return detect, bg_cutoff
-
-def get_best_fit_seds(morphs, images):
-    """Calculate best fitting SED for multiple components.
-
-    Solves min_A ||img - AS||^2 for the SED matrix A,
-    assuming that the images only contain a single source.
-
-    Parameters
-    ----------
-    morphs: list
-        Morphology for each component in the source.
-    frame: `scarlet.observation.frame`
-        The frame of the model
-    images: array
-        Observation to extract SEDs from.
-
-    Returns
-    -------
-    SED: `~numpy.array`
-    """
-    K = len(morphs)
-    _morph = morphs.reshape(K, -1)
-    data = images.reshape(images.shape[0], -1)
-    seds = np.dot(np.linalg.inv(np.dot(_morph, _morph.T)), np.dot(_morph, data.T))
-    return seds
-
-def trim_morphology(sky_coord, frame, morph, bg_cutoff, thresh):
-    # trim morph to pixels above threshold
-    mask = morph > bg_cutoff * thresh
-    boxsize = 16
-    pixel_center = frame.get_pixel(sky_coord)
-    if mask.sum() > 0:
-        morph[~mask] = 0
-
-        # normalize to unity at peak pixel
-        center_morph = morph[pixel_center[0], pixel_center[1]]
-        morph /= center_morph
-
-        # find fitting bbox
-        bbox = Box.from_data(morph, min_value=0)
-        if bbox.contains(pixel_center):
-            size = 2 * max(
-                (
-                    pixel_center[0] - bbox.start[-2],
-                    bbox.stop[0] - pixel_center[-2],
-                    pixel_center[1] - bbox.start[-1],
-                    bbox.stop[1] - pixel_center[-1],
-                )
-            )
-            while boxsize < size:
-                boxsize *= 2
-    else:
-        msg = "No flux above threshold for source at y={0} x={1}".format(*pixel_center)
-        logger.warning(msg)
-
-    # define bbox and trim to bbox
-    bottom = pixel_center[0] - boxsize // 2
-    top = pixel_center[0] + boxsize // 2
-    left = pixel_center[1] - boxsize // 2
-    right = pixel_center[1] + boxsize // 2
-    bbox = Box.from_bounds((bottom, top), (left, right))
-    morph = bbox.extract_from(morph)
-    bbox_3d = Box.from_bounds((0, frame.C), (bottom, top), (left, right))
-    return morph, bbox_3d
-
-
-def init_extended_source(
-    sky_coord,
-    frame,
-    observations,
-    coadd = None,
-    bg_cutoff=None,
-    thresh=1,
-    symmetric=True,
-    monotonic=True,
-    min_grad=0.2,
-):
-    """Initialize the source that is symmetric and monotonic
-    See `ExtendedSource` for a description of the parameters
-    """
-    try:
-        iter(observations)
-    except TypeError:
-        observations = [observations]
-    # determine initial SED from peak position
-    # SED in the frame for source detection
-    seds = []
-    for obs in observations:
-        _sed = get_psf_sed(sky_coord, obs, frame)
-        seds.append(_sed)
-    sed = np.concatenate(seds).flatten()
-
-    if np.all(sed <= 0):
-        # If the flux in all channels is  <=0,
-        msg = f"Zero or negative SED {sed} at y={sky_coord[0]}, x={sky_coord[1]}"
-        logger.warning(msg)
-    if coadd is None:
-        # which observation to use for detection and morphology
-        try:
-            bg_rms = np.array([[1 / np.sqrt(w[w > 0].mean()) for w in obs_.weights] for obs_ in observations])
-        except:
-            raise AttributeError(
-                "Observation.weights missing! Please set inverse variance weights"
-            )
-        coadd, bg_cutoff = build_sed_coadd(seds, bg_rms, observations)
-    else:
-        if bg_cutoff is None:
-            raise AttributeError(
-                "background cutoff missing! Please set argument bg_cutoff"
-            )
-    # Apply the necessary constraints
-    center = frame.get_pixel(sky_coord)
-    if symmetric:
-        morph = operator.prox_uncentered_symmetry(
-            coadd*1., 0, center=center, algorithm="sdss" # *1 is to artificially pass a variable that is not coadd
-        )
-    else:
-        morph = coadd
-    if monotonic:
-        # use finite thresh to remove flat bridges
-        prox_monotonic = operator.prox_weighted_monotonic(
-            morph.shape, neighbor_weight="flat", center=center, min_gradient=min_grad
-        )
-        morph = prox_monotonic(morph, 0).reshape(morph.shape)
-
-    morph, bbox = trim_morphology(sky_coord, frame, morph, bg_cutoff, thresh)
-    return sed, morph, bbox
-
-
-def init_multicomponent_source(
-    sky_coord,
-    frame,
-    observations,
-    coadd = None,
-    bg_cutoff = None,
-    flux_percentiles=None,
-    thresh=1,
-    symmetric=True,
-    monotonic=True,
-    min_grad=0.2,
-):
-    """Initialize multiple components
-    See `MultiComponentSource` for a description of the parameters
-    """
-    try:
-        iter(observations)
-    except TypeError:
-        observations = [observations]
-
-    # The observation that lives in the same plane as the frame
-    loc = np.where([type(obs) is Observation for obs in observations])
-    # If more than one element is an `Observation`, then pick the first one as a reference (arbitrary)
-    obs_ref = observations[loc[0][0]]
-
-    if flux_percentiles is None:
-        flux_percentiles = [25]
-
-    # Initialize the first component as an extended source
-    sed, morph, bbox = init_extended_source(
-        sky_coord,
-        frame,
-        observations,
-        coadd=coadd,
-        bg_cutoff=bg_cutoff,
-        thresh=thresh,
-        symmetric=symmetric,
-        monotonic=monotonic,
-        min_grad=min_grad
-    )
-    # create a list of components from base morph by layering them on top of
-    # each other so that they sum up to morph
-    K = len(flux_percentiles) + 1
-
-    Ny, Nx = morph.shape
-    morphs = np.zeros((K, Ny, Nx), dtype=morph.dtype)
-    morphs[0, :, :] = morph[:, :]
-    max_flux = morph.max()
-    percentiles_ = np.sort(flux_percentiles)
-    last_thresh = 0
-    for k in range(1, K):
-        perc = percentiles_[k - 1]
-        flux_thresh = perc * max_flux / 100
-        mask_ = morph > flux_thresh
-        morphs[k - 1][mask_] = flux_thresh - last_thresh
-        morphs[k][mask_] = morph[mask_] - flux_thresh
-        last_thresh = flux_thresh
-
-    # renormalize morphs: initially Smax
-    for k in range(K):
-        if np.all(morphs[k] <= 0):
-            msg = "Zero or negative morphology for component {} at y={}, x={}"
-            logger.warning(msg.format(k, *sky_coord))
-        morphs[k] /= morphs[k].max()
-
-    # optimal SEDs given the morphologies, assuming img only has that source
-    boxed_img = bbox.extract_from(obs_ref.images)
-    seds = get_best_fit_seds(morphs, boxed_img)
-
-    for k in range(K):
-        if np.all(seds[k] <= 0):
-            # If the flux in all channels is  <=0,
-            # the new sed will be filled with NaN values,
-            # which will cause the code to crash later
-            msg = "Zero or negative SED {} for component {} at y={}, x={}".format(
-                seds[k], k, *sky_coord
-            )
-            logger.warning(msg)
-
-    return seds, morphs, bbox
-
-class Random:
-    """ Class used to instantiate a RandomSource class from its kwargs.
-    """
-    def __init__(self, observations = None):
-        """Sets a *Source with all its arguments"""
-        self.kwargs = observations
-    def __call__(self, *args):
-        """Sets a *Source with all its arguments"""
-        return RandomSource(args[0], observation = self.kwargs)
 
 class RandomSource(FactorizedComponent):
     """Sources with uniform random morphology and sed.
@@ -381,12 +51,6 @@ class RandomSource(FactorizedComponent):
         )
 
         super().__init__(model_frame, model_frame.bbox, sed, morph)
-
-class Point:
-    """ Class used to instantiate a PointSource class from its kwargs.
-    """
-    def __call__(self, *args):
-        return PointSource(*args)
 
 class PointSource(FunctionComponent):
     """Source intialized with a single pixel
@@ -456,29 +120,6 @@ class PointSource(FunctionComponent):
 
     def _psf_wrapper(self, *parameters):
         return self.model_frame.psf.__call__(*parameters, bbox=self.bbox)[0]
-
-class Starlets:
-    """ "Setter" class to initialise a `StarletSource` from its keyword arguments.
-
-        Attributes
-        ----------
-        thresh: `float`
-            Multiple of the backround RMS used as a
-            flux cutoff for morphology initialization.
-        starlet_thresh: `float`
-            threshold for wavelet coefficients in units of the noise std.
-    """
-    def __init__(
-        self,
-        thresh=1.0,
-        starlet_thresh=5,
-        min_grad = 0,
-    ):
-        self.kwargs = (thresh,  starlet_thresh, min_grad)
-
-    def __call__(self, *args, coadd = None, bg_cutoff = None):
-        """Sets a *Source with all its arguments"""
-        return StarletSource(*args, coadd, bg_cutoff, *self.kwargs)
 
 class StarletSource(FunctionComponent):
     """Source intialized with starlet coefficients.
@@ -584,37 +225,6 @@ class StarletSource(FunctionComponent):
         """
         return Starlet(coefficients = param).image[0]
 
-class Extended:
-    """ "Setter" class to initialise an `ExtendedSource` from its keyword arguments only
-
-        Attributes
-        ----------
-        thresh: `float`
-            Multiple of the backround RMS used as a
-            flux cutoff for morphology initialization.
-        monotonic: ['flat', 'angle', 'nearest'] or None
-            Which version of monotonic decrease in flux from the center to enforce
-        symmetric: `bool`
-            Whether or not to enforce symmetry.
-        shifting: `bool`
-            Whether or not a subpixel shift is added as optimization parameter
-        min_grad: `float`
-            sets the "strength" of the monotonicity operator (default = 0.2 is strong!)
-    """
-    def __init__(
-        self,
-        thresh=1.0,
-        monotonic="flat",
-        symmetric=False,
-        shifting=False,
-        min_grad = 0.2
-    ):
-        self.kwargs = (thresh,  monotonic,  symmetric, shifting, min_grad)
-
-    def __call__(self, *args, coadd = None, bg_cutoff = None):
-        """Sets a *Source with all its arguments"""
-        return ExtendedSource(*args, coadd, bg_cutoff, *self.kwargs)
-
 class ExtendedSource(FactorizedComponent):
     def __init__(
         self,
@@ -627,7 +237,7 @@ class ExtendedSource(FactorizedComponent):
         monotonic="flat",
         symmetric=False,
         shifting=False,
-        min_grad = 0.2,
+        min_grad=0.2,
     ):
         """Extended source intialized to match a set of observations
 
@@ -717,39 +327,6 @@ class ExtendedSource(FactorizedComponent):
             return self.pixel_center + self.shift
         else:
             return self.pixel_center
-
-class MultiComponent:
-    """ "Setter" class to initialise an `ExtendedSource` from its keyword arguments only
-
-            Attributes
-            ----------
-            thresh: `float`
-                Multiple of the backround RMS used as a
-                flux cutoff for morphology initialization.
-            monotonic: ['flat', 'angle', 'nearest'] or None
-                Which version of monotonic decrease in flux from the center to enforce
-            symmetric: `bool`
-                Whether or not to enforce symmetry.
-            shifting: `bool`
-                Whether or not a subpixel shift is added as optimization parameter
-            min_grad: `float`
-                sets the "strength" of the monotonicity operator (default = 0.2 is strong!)
-        """
-
-    def __init__(
-            self,
-            thresh=1.0,
-            flux_percentiles=None,
-            monotonic="flat",
-            symmetric=False,
-            shifting=False,
-            min_grad=0.2
-    ):
-        self.kwargs = (thresh, flux_percentiles, symmetric, monotonic, shifting, min_grad)
-
-    def __call__(self, *args, coadd=None, bg_cutoff=None):
-        """Sets a *Source with all its arguments"""
-        return MultiComponentSource(*args, coadd, bg_cutoff, *self.kwargs)
 
 class MultiComponentSource(ComponentTree):
     """Extended source with multiple components layered vertically.
